Accurate stick recognition
recognize the angle of the stick as the true signifier that a shot
should be fired.

fast rotation can still cause the shot to be missed.

need to get the shot to go in the proper angle, but this finally uses
the stick to determine if a 'shot' should be applied.